# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HyperLocal Business Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/businesses/{businessId}/contacts:
    get:
      operationId: listBusinessContacts
      tags: [businesses]
      summary: List contacts for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BusinessContact"
    post:
      operationId: createBusinessContact
      tags: [businesses]
      summary: Add a contact to a business (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessContactInput"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BusinessContact"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /v1/businesses/{businessId}/contacts/{contactId}:
    patch:
      operationId: updateBusinessContact
      tags: [businesses]
      summary: Update a business contact (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessContactInput"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BusinessContact"
    delete:
      operationId: deleteBusinessContact
      tags: [businesses]
      summary: Delete a business contact (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Contact deleted

  # ── Business Branches ─────────────────────────────────────────────────────────
...
  /v1/search/updates:
    get:
      operationId: searchUpdates
      tags: [search]
      summary: Search published business updates (news, offers, events, announcements)
      parameters:
        - name: q
          in: query
          description: Full-text search query for update titles
          schema:
            type: string
        - name: businessId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          description: Filter by update type
          schema:
            type: string
            enum: [news, offer, event, announcement]
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of business updates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        businessId:
                          type: string
                          format: uuid
                        type:
                          type: string
                          enum: [news, offer, event, announcement]
                        title:
                          type: string
                        body:
                          type: string
                        businessName:
                          type: string
                        businessSlug:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                  nextCursor:
                    type: string
                    nullable: true
                  total:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /v1/businesses/{businessId}/updates:
    get:
      operationId: listBusinessUpdates
      tags: [business-updates]
      summary: List published updates for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of business updates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BusinessUpdateItem"
    post:
      operationId: createBusinessUpdate
      tags: [business-updates]
      summary: Owner — create a business update
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title:
                  type: string
                  maxLength: 200
                body:
                  type: string
                  maxLength: 5000
                updateType:
                  type: string
                  enum: [news, offer, event, announcement]
                  default: announcement
                publishNow:
                  type: boolean
                  default: true
                expiresAt:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Update created
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /v1/businesses/{businessId}/updates/{updateId}:
    patch:
      operationId: patchBusinessUpdate
      tags: [business-updates]
      summary: Owner — edit a business update
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: updateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                updateType:
                  type: string
                  enum: [news, offer, event, announcement]
                status:
                  type: string
                  enum: [draft, published, archived]
      responses:
        "200":
          description: Update patched
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    delete:
      operationId: deleteBusinessUpdate
      tags: [business-updates]
      summary: Owner — soft-delete a business update
      security:
        - bearerAuth: []
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: updateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: marketplace
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Update deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    BusinessContact:
      type: object
      required: [id, businessId, contactType, value]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
        contactType:
          type: string
        value:
          type: string
        isPrimary:
          type: boolean
        isVerified:
          type: boolean
        displayOrder:
          type: integer
    BusinessContactInput:
      type: object
      required: [contactType, value]
      properties:
        contactType:
          type: string
          enum: [phone, whatsapp, email, website, instagram, facebook, twitter, tiktok, youtube, other]
        value:
          type: string
        isPrimary:
          type: boolean
        displayOrder:
          type: integer
...
    BusinessUpdateItem:
      type: object
      required: [id, businessId, title, body, updateType, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        updateType:
          type: string
          enum: [news, offer, event, announcement]
        status:
          type: string
        publishedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time